improved compiler and minijava implementation

+++ plzoo-workspace/org.tetrabox.minijava.lang.interpreter2/interpreter-comp/miniJava/interpreter/miniJava/impl/ModuloImpl.java
@@ -161,7 +161,7 @@ public class ModuloImpl extends ExpressionImpl implements Modulo {
 				IntegerValue bleft = ((IntegerValue) (left));
 				IntegerValue bright = ((IntegerValue) (right));
 				IntegerValue tmp = ((IntegerValue) (MiniJavaFactory.eINSTANCE.createIntegerValue()));
-				tmp.setValue(MathService.mod((Modulo) (this), (bleft.getValue()), (bright.getValue())));
+				tmp.setValue(MathService.mod((Modulo) (this), (int) (bleft.getValue()), (int) (bright.getValue())));
 				result = (Value) (tmp) ;
 			}
 			else {
