+++ plzoo-workspace/org.tetrabox.minijava.lang.interpreter2/revisitor-comp/minijava_exec/impl/operation/impl/ModuloOpImpl.java
@@ -107,7 +107,7 @@ public class ModuloOpImpl extends ExpressionOpImpl implements ModuloOp {
         IntegerValue bleft = ((IntegerValue) (left));
         IntegerValue bright = ((IntegerValue) (right));
         IntegerValue tmp = ((IntegerValue) (MiniJavaFactory.eINSTANCE.createIntegerValue()));
-        tmp.setValue(MathService.mod((Modulo) (this.obj), (bleft.getValue()), (bright.getValue())));
+        tmp.setValue(MathService.mod((Modulo) (this.obj), (int) (bleft.getValue()), (int) (bright.getValue())));
         result = tmp;
       }
       else {
